+++ cmd/status.go
@@ -21,7 +21,7 @@ func NewStatusCommand() *cobra.Command {
 }
 
 func handleStatus() error {
-	vault, err := storage.LoadVault()
+	vault, _, err := storage.LoadVault()
 	if err != nil {
 		return err
 	}